chore: remove older next.js demo
