+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
 	"eslint.packageManager": "pnpm",
-	"eslint.workingDirectories": ["./packages/lib", "./demos/nextjs", "./demos/nextjs-app-router"],
+	"eslint.workingDirectories": ["./packages/lib", "./demos/nextjs"],
 	"editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   },
