+++ demos/nextjs-app-router/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "nextjs-app-router-demo",
+	"name": "nextjs-demo",
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
