+++ demos/nextjs-app-router/src/app/demo/page.tsx
@@ -1,0 +1,10 @@
+import Demo from '~/components/Demo'
+
+export default function Page() {
+	return (
+		<>
+			<h1 className="mb-4 text-4xl font-semibold">React Turnstile Demo</h1>
+			<Demo />
+		</>
+	)
+}
