--- demos/nextjs-app-router/src/app/page.tsx
@@ -1,11 +1,0 @@
-export default function Home() {
-	return (
-		<main className="bg-[#1d1f20]">
-			<div className="flex min-h-screen w-full flex-col items-center justify-center py-24">
-				<main className="flex w-full max-w-[740px] flex-col justify-center gap-2 p-4 text-white">
-					<h1 className="mb-4 text-4xl font-semibold">React Turnstile Demo</h1>
-				</main>
-			</div>
-		</main>
-	)
-}
