+++ demos/nextjs-app-router/src/components/ConfigForm.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, useState } from 'react'
-import { langOptions, siteKeyOptions, sizeOptions, themeOptions } from '../constants'
-import { SiteKeyType, WidgetSize } from '../types'
+import { SiteKeyType, WidgetSize } from '~/types'
+import { langOptions, siteKeyOptions, sizeOptions, themeOptions } from '~/constants'
 import Options from './Options'
 
 interface FormProps {
