+++ demos/nextjs-app-router/src/components/Footer.tsx
