+++ demos/nextjs-app-router/src/components/WidgetMethods.tsx
