+++ demos/nextjs-app-router/src/constants.ts
