+++ demos/nextjs-app-router/tsconfig.json
@@ -20,7 +20,7 @@
 			}
 		],
 		"paths": {
-			"@/*": ["./src/*"]
+			"~/*": ["./src/*"]
 		}
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
