--- demos/nextjs/.prettierrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-	...require('../../.prettierrc.js'),
-	plugins: [require('prettier-plugin-tailwindcss')]
-}
