--- demos/nextjs/next.config.js
@@ -1,7 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-	reactStrictMode: true,
-	swcMinify: true
-}
-
-module.exports = nextConfig
