--- demos/nextjs/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "nextjs-demo",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"dev": "next dev -p 3000",
-		"build": "pnpm run --filter=@marsidev/react-turnstile build && next build",
-		"start": "next start",
-		"lint": "next lint",
-		"lint:fix": "next lint --fix",
-		"type-check": "tsc --noEmit"
-	},
-	"dependencies": {
-		"@marsidev/react-turnstile": "workspace:*",
-		"classnames": "2.3.2",
-		"next": "13.2.3",
-		"react": "18.2.0",
-		"react-dom": "18.2.0"
-	},
-	"devDependencies": {
-		"autoprefixer": "10.4.13",
-		"eslint-config-custom": "workspace:*",
-		"eslint-config-next": "13.2.3",
-		"postcss": "8.4.21",
-		"prettier-plugin-tailwindcss": "0.2.4",
-		"tailwindcss": "3.2.7"
-	},
-	"eslintConfig": {
-		"extends": [
-			"next/core-web-vitals",
-			"custom"
-		]
-	}
-}
