--- demos/nextjs/src/styles/globals.css
@@ -1,7 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-html {
-	color-scheme: dark;
-}
